Custom Trigger Volume script for Unity.
I am particularly proud of this one. I researched and learned how to make custom trigger volumes that any of our engineers can use to do anything they need in Unity. Although our game does not need it right, it would be easy to add the ability to have the trigger volume trigger sounds and music.